knn and its visualize

+++ data-proc/normalize_all_sliver07.sh
@@ -1,20 +1,22 @@
 #!/bin/bash
 # $File: normalize_all_sliver07.sh
-# $Date: Mon May 11 19:56:20 2015 +0800
+# $Date: Mon May 11 21:17:41 2015 +0800
 # $Author: jiakai <jia.kai66@gmail.com>
 
 cropper=$(realpath affine_normalize_img.py)
 border_dist=$(realpath border_dist_tools/calc_border_dist.py)
 
-set +e
-cd ../../sliver07/train-orig
+if ! cd ../../sliver07/train-orig
+then
+    echo "training data do not exist"
+    exit -1
+fi
+
 rm -rf ../train-cropped
 rm -rf ../border-dist
 
 mkdir ../train-cropped
 mkdir ../border-dist
-
-set -e
 
 for i in $(seq 1 20)
 do
@@ -23,6 +25,6 @@ do
     output=../train-cropped/$outnum
     $cropper -i $(printf 'liver-orig%03d.nii.gz' $i) \
         --mask $(printf 'liver-seg%03d.nii.gz' $i) \
-        -o $output --scale 0.8
+        -o $output --scale 0.8 --min_mask_border 11
     $border_dist ${output}-mask.nii.gz ../border-dist/${outnum}.nii.gz
 done
